// src/models/dataModel.ts
// Generated by https://quicktype.io

export interface RespDataJSON {
  categorias: Categoria[];
}

export interface Categoria {
  categoria: string;
  productos: Producto[];
}

export interface Producto {
  producto: string;
  marcas: Marca[];
}

export interface Marca {
  marca: string;
  ventas: number[];
}